Created simple brain class that thinks.
Created Brain class that as of now takes five simple attributes from the env / organism and returns a vector of values that will help our organism move. Added new method to organism class in order to translate the sensory attributes into numpy arrays for calculation, which only needs to be done at the initialization of the genome of the class. Would like to look into whether or not I can speed up the thinking process.