Created functioning brain. Organisms now evolve and (eventually) move to safe area.
Accidentally a huge commit. Organisms "learn" generation to generation to move to the safe area. Thinking calculations made using numpy, the genome of the organism is converted to a matrix on initialization. Animation still slow, most likely due to thinking and the fact that we need to loop through each organism in order to calculate their next move. Looking forward to speeding it up and organizing the coupling / cohesion of the classes more.